Move BirdtrayEventFilter to a separate file
