+++ CMakeLists.txt
@@ -21,8 +21,8 @@ if (WIN32)
     add_definitions(-DUNICODE)
     find_library(USER32 user32)
     list(APPEND REQUIRED_MODULES ${USER32})
-    set(PLATFORM_SOURCES src/windowtools_win.cpp src/res/birdtray.rc)
-    set(PLATFORM_HEADERS src/windowtools_win.h)
+    set(PLATFORM_SOURCES src/windowtools_win.cpp src/birdtrayeventfilter.cpp src/res/birdtray.rc)
+    set(PLATFORM_HEADERS src/windowtools_win.h src/birdtrayeventfilter.h)
 else()
     find_package(Qt5X11Extras QUIET)
     find_library(X11 X11)
